Refactoring: rename beans to config
