--- src/main/java/tech/jhipster/light/generator/server/javatool/base/infrastructure/beans/JavaBaseBeanConfiguration.java
@@ -1,22 +1,0 @@
-package tech.jhipster.light.generator.server.javatool.base.infrastructure.beans;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import tech.jhipster.light.generator.project.domain.ProjectRepository;
-import tech.jhipster.light.generator.server.javatool.base.domain.JavaBaseDomainService;
-import tech.jhipster.light.generator.server.javatool.base.domain.JavaBaseService;
-
-@Configuration
-public class JavaBaseBeanConfiguration {
-
-  private final ProjectRepository projectRepository;
-
-  public JavaBaseBeanConfiguration(ProjectRepository projectRepository) {
-    this.projectRepository = projectRepository;
-  }
-
-  @Bean
-  public JavaBaseService javaBaseService() {
-    return new JavaBaseDomainService(projectRepository);
-  }
-}
