--- src/main/java/tech/jhipster/light/generator/server/springboot/core/infrastructure/beans/SpringBootServiceBeanConfiguration.java
@@ -1,25 +1,0 @@
-package tech.jhipster.light.generator.server.springboot.core.infrastructure.beans;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import tech.jhipster.light.generator.buildtool.generic.domain.BuildToolService;
-import tech.jhipster.light.generator.project.domain.ProjectRepository;
-import tech.jhipster.light.generator.server.springboot.core.domain.SpringBootDomainService;
-import tech.jhipster.light.generator.server.springboot.core.domain.SpringBootService;
-
-@Configuration
-public class SpringBootServiceBeanConfiguration {
-
-  private final ProjectRepository projectRepository;
-  private final BuildToolService buildToolService;
-
-  public SpringBootServiceBeanConfiguration(ProjectRepository projectRepository, BuildToolService buildToolService) {
-    this.projectRepository = projectRepository;
-    this.buildToolService = buildToolService;
-  }
-
-  @Bean
-  public SpringBootService springBootService() {
-    return new SpringBootDomainService(projectRepository, buildToolService);
-  }
-}
