--- src/main/java/tech/jhipster/light/generator/server/springboot/dbmigration/liquibase/infrastructure/beans/LiquibaseServiceBeanConfiguration.java
@@ -1,32 +1,0 @@
-package tech.jhipster.light.generator.server.springboot.dbmigration.liquibase.infrastructure.beans;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import tech.jhipster.light.generator.buildtool.generic.domain.BuildToolService;
-import tech.jhipster.light.generator.project.domain.ProjectRepository;
-import tech.jhipster.light.generator.server.springboot.dbmigration.liquibase.domain.LiquibaseDomainService;
-import tech.jhipster.light.generator.server.springboot.dbmigration.liquibase.domain.LiquibaseService;
-import tech.jhipster.light.generator.server.springboot.properties.domain.SpringBootPropertiesService;
-
-@Configuration
-public class LiquibaseServiceBeanConfiguration {
-
-  public final ProjectRepository projectRepository;
-  public final BuildToolService buildToolService;
-  public final SpringBootPropertiesService springBootPropertiesService;
-
-  public LiquibaseServiceBeanConfiguration(
-    ProjectRepository projectRepository,
-    BuildToolService buildToolService,
-    SpringBootPropertiesService springBootPropertiesService
-  ) {
-    this.projectRepository = projectRepository;
-    this.buildToolService = buildToolService;
-    this.springBootPropertiesService = springBootPropertiesService;
-  }
-
-  @Bean
-  public LiquibaseService liquibaseService() {
-    return new LiquibaseDomainService(projectRepository, buildToolService, springBootPropertiesService);
-  }
-}
